[resurrect] Allow users to register minimum functions for custom datatypes
This commit allows users to register global functions named
`tvm.datatype.min.<type name>` which take the number of bits in the custom type
and return the corresponding minimum value (as a double).

A similar commit will need to be created for max, whenever that ends up being
needed!

This commit was resurrected, after I threw it out accidentally in a bad merge.